added a reminder when the project dir already exits

+++ src/utility.py
@@ -46,7 +46,7 @@ class checkpoint():
         self.args = args
         self.ok = True
         self.log = torch.Tensor()
-        now = datetime.datetime.now().strftime('%Y-%m-%d-%H:%M:%S')
+        now = datetime.datetime.now().strftime('%Y%m%d_%H%M%S')
 
         if not args.load:
             if not args.save:
@@ -59,6 +59,17 @@ class checkpoint():
                 print('Continue from epoch {}...'.format(len(self.log)))
             else:
                 args.load = ''
+
+        # add a reminder
+        if os.path.exists(self.dir):
+            val = input('Warning: directory "%s" already exists, is there any potential problem with this? Type [yes/no] to continue: ' % self.dir)
+            if val.lower() == 'no':
+                val = input('Warning: Are you sure? We cannot be too careful. Type [yes/no] to continue: ')
+            if val.lower() == 'yes':
+                print("You've responded with 'yes'. This program is about to terminate. Please check and run again.")
+                exit(1)
+            else:
+                print("You are very positive that there is NO problem. The program continues running. Have a nice day!")
 
         if args.reset:
             os.system('rm -rf ' + self.dir)
